backend/server/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.1
daphne==4.0.0
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.7
pillow==10.2.0
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
zope.interface==6.1
